refactor: structural decomposition for cohesion and maintainability
This PR improves the internal structure of isbn3 without changing any
public API behavior. All existing tests pass (66/66) and a regression
test suite (Regrets) confirms zero behavioral changes.

## Changes

### Decomposition
- **audit.js** (118 → 30 lines): Extracted clue-generating helpers to
  `lib/audit_clues.js`. Each function now has a single responsibility
  and JSDoc documentation.

### Cohesion — Inline functions → dedicated modules
- **hyphenate**: Extracted from inline arrow function in isbn.js to
  `lib/hyphenate.js`
- **asIsbn13**: Extracted from inline arrow function in isbn.js to
  `lib/as_isbn13.js`
- **asIsbn10**: Extracted from inline arrow function in isbn.js to
  `lib/as_isbn10.js`

### Naming — Honest names
- **fill.js → complete_isbn_data.js**: The old name was vague. The new
  name clearly describes that this module completes ISBN data objects
  with computed check digits and formatted versions.

### Single Responsibility
- **normalize**: Extracted from audit.js private function to
  `lib/normalize.js` — useful as a standalone utility.

### Bug Fix (Typo)
- `lookForPossibleInvalityCauses` →
  `lookForPossibleInvalidityCauses`

## Verification

All changes were verified with three independent methods:

1. **Regrets regression testing**: 5 clusters, all GREEN, zero drift
2. **Direct output comparison**: All 15 function outputs identical to
   pre-refactor baseline
3. **Fingerprint cross-match**: All 5 fingerprints match pre-refactor
   contract
