+++ isbn.js
@@ -1,28 +1,18 @@
+// isbn.js — Main entry point for the isbn3 library
+// Refactored: inline functions extracted to dedicated modules for cohesion
+
 const parse = require('./lib/parse')
 const audit = require('./lib/audit')
+const hyphenate = require('./lib/hyphenate')
+const asIsbn13 = require('./lib/as_isbn13')
+const asIsbn10 = require('./lib/as_isbn10')
 
 module.exports = {
   parse,
   audit,
-
-  hyphenate: val => {
-    const data = parse(val)
-    if (!data) return null
-    return data.isIsbn13 ? data.isbn13h : data.isbn10h
-  },
-  asIsbn13: (val, hyphen) => {
-    const data = parse(val)
-    if (!data) return null
-    return hyphen ? data.isbn13h : data.isbn13
-  },
-  asIsbn10: (val, hyphen) => {
-    const data = parse(val)
-    if (!data) return null
-    // Return null for cases where it shouldn't map to an ISBN 10
-    // ex: 979-10-91146-13-5
-    if (!data.isbn10) return null
-    return hyphen ? data.isbn10h : data.isbn10
-  },
+  hyphenate,
+  asIsbn13,
+  asIsbn10,
 
   groups: require('./lib/groups')
 }
