+++ lib/parse.js
@@ -1,4 +1,7 @@
-const fill = require('./fill')
+// parse.js — Parse an ISBN string and return structured data
+// Updated to use renamed complete_isbn_data module
+
+const completeIsbnData = require('./complete_isbn_data')
 const splitIsbnParts = require('./split_isbn_parts')
 
 module.exports = value => {
@@ -27,7 +30,7 @@ module.exports = value => {
     data.isIsbn13 = false
   }
 
-  data = fill(data)
+  data = completeIsbnData(data)
   if (!data) return null
 
   data.isValid = data.check === (data.isIsbn13 ? data.check13 : data.check10)
